requirements.txt
asgiref==3.6.0
Django==4.1.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
psycopg2==2.9.5
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
